Rename namespace + fix CRLF to LF
