+++ OpenKh.Kh2/SystemData/Item.cs
@@ -1,8 +1,8 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.IO;
 using Xe.BinaryMapper;
 
-namespace OpenKh.Kh2.System
+namespace OpenKh.Kh2.SystemData
 {
     public class Item
     {
