+++ OpenKh.Kh2/SystemData/Trsr.cs
@@ -1,9 +1,8 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using Xe.BinaryMapper;
 
-namespace OpenKh.Kh2.System
+namespace OpenKh.Kh2.SystemData
 {
     public class Trsr
     {
