+++ OpenKh.Tools.Kh2SystemEditor/Models/Export/ItemExport.cs
@@ -1,5 +1,5 @@
-﻿using OpenKh.Engine;
-using OpenKh.Kh2.System;
+using OpenKh.Engine;
+using OpenKh.Kh2.SystemData;
 using OpenKh.Tools.Kh2SystemEditor.Attributes;
 using System;
 using System.Collections.Generic;
